requirements.txt
playwright>=1.40.0
python-docx>=1.1.0
PyYAML>=6.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
pypdf>=4.0.0
loguru>=0.7.0
tenacity>=8.2.0
fake-useragent>=1.4.0
openai>=1.0.0
